Defend Rewritten, Bug Report
The Defend state has been rewritten using our new event driven style. There is a bug when the player dies where the hurt effect doesn't play out correctly, also the player spawns faded as if they've just been hurt. This is likely due to our change in persistence when the player respawns. We'll have to revert that change.